created custom exception class and implemented in Doctor's service layer

+++ src/main/java/com/markfox/patientmanager/services/impls/DoctorServiceImpl.java
@@ -1,5 +1,6 @@
 package com.markfox.patientmanager.services.impls;
 
+import com.markfox.patientmanager.exceptions.MyException;
 import com.markfox.patientmanager.models.Doctor;
 import com.markfox.patientmanager.models.Patient;
 import com.markfox.patientmanager.repositories.DoctorRepository;
@@ -24,22 +25,34 @@ public class DoctorServiceImpl implements DoctorService {
     }
 
     @Override
-    public Doctor getDoctorById(Long id) {
+    public Doctor getDoctorById(Long id) throws MyException {
+        if (id == null || id < 1) {
+            throw new MyException("Error: Doctor ID cannot be null");
+        }
         return doctorRepository.findById(id).isPresent() ? doctorRepository.findById(id).get() : null;
     }
 
     @Override
-    public List<Patient> getAllDocsPatients(Long id) {
+    public List<Patient> getAllDocsPatients(Long id) throws MyException {
+        if (id == null || id < 1) {
+            throw new MyException("Error: Doctor ID cannot be null");
+        }
         return doctorRepository.findById(id).isPresent() ? doctorRepository.findById(id).get().getDocsPatients() : null;
     }
 
     @Override
-    public Doctor addDoctor(Doctor doctor) {
+    public Doctor addDoctor(Doctor doctor) throws MyException {
+        if (doctor == null) {
+            throw new MyException("Error: Doctor object cannot be null");
+        }
         return doctorRepository.save(doctor);
     }
 
     @Override
-    public void deleteDoctorById(Long id) {
+    public void deleteDoctorById(Long id) throws MyException {
+        if (id == null || id < 1) {
+            throw new MyException("Error: Doctor ID cannot be null");
+        }
         doctorRepository.deleteById(id);
     }
 }
